medgemma-api/requirements.txt
torch
Pillow
huggingface_hub
transformers
bitsandbytes
fastapi
uvicorn
python-multipart